Pull drawable classes out of gmplot.py.
