+++ gmplot/gmplot.py
@@ -5,14 +5,20 @@ import math
 import os
 import requests
 import warnings
-import math
 
 from collections import namedtuple
 
 from gmplot.color import _get_hex_color
-from gmplot.google_maps_templates import SYMBOLS, CIRCLE_MARKER
-from gmplot.utility import StringIO, _get_value, _get_embeddable_image, _INDENT_LEVEL
+from gmplot.google_maps_templates import SYMBOLS, CIRCLE_MARKER # TODO: Remove CIRCLE_MARKER (counts as an API change).
+from gmplot.utility import _INDENT_LEVEL, StringIO, _get_value, _format_LatLng
 from gmplot.writer import _Writer
+
+from gmplot.drawables.marker_dropper import _MarkerDropper
+from gmplot.drawables.marker_icon import _MarkerIcon
+from gmplot.drawables.marker_info_window import _MarkerInfoWindow
+from gmplot.drawables.marker import _Marker
+from gmplot.drawables.route import _Route
+from gmplot.drawables.text import _Text
 
 Symbol = namedtuple('Symbol', ['symbol', 'lat', 'long', 'size'])
 # TODO: Rename `long` to `lng` to match the rest of the project (counts as an API change).
@@ -30,290 +36,6 @@ def safe_iter(var): # TODO: Remove since unused (counts as API change).
         return iter(var)
     except TypeError:
         return [var]
-
-_COLOR_ICON_PATH = os.path.join(os.path.dirname(__file__), 'markers/%s.png')
-
-def _format_LatLng(lat, lng, precision=6): # TODO: Make static method of GoogleMapPlotter.
-    '''
-    Format the given latitude/longitude location as a Google Maps LatLng object.
-
-    Args:
-        lat (float): Latitude.
-        lng (float): Longitude.
-
-    Optional:
-
-    Args:
-        precision (int): Number of digits after the decimal to round to for lat/lng values. Defaults to 6.
-
-    Returns:
-        str: Formatted Google Maps LatLng object.
-    '''
-    return 'new google.maps.LatLng(%.*f, %.*f)' % (precision, lat, precision, lng)
-
-class _Route(object):
-    '''For more info, see Google Maps' `Directions Service https://developers.google.com/maps/documentation/javascript/directions`_.'''
-    
-    def __init__(self, origin, destination, **kwargs):
-        '''
-        Args:
-            origin ((float, float)): Origin, as a latitude/longitude tuple.
-            destination ((float, float)): Destination, as a latitude/longitude tuple.
-
-        Optional:
-
-        Args:
-            travel_mode (str): Travel mode. Defaults to 'DRIVING'.
-            waypoints ([(float, float)]): Waypoints.
-        '''
-        self._origin = _format_LatLng(*origin)
-        self._destination = _format_LatLng(*destination)
-        self._travel_mode = _get_value(kwargs, ['travel_mode'], 'DRIVING').upper()
-        self._waypoints = [_format_LatLng(*waypoint) for waypoint in _get_value(kwargs, ['waypoints'], [])]
-
-    def write(self, w):
-        '''
-        Write the route.
-
-        Args:
-            w (_Writer): Writer used to write the route.
-        '''
-        w.write('new google.maps.DirectionsService().route({')
-        w.indent()
-        w.write('origin: %s,' % self._origin)
-        w.write('destination: %s,' % self._destination)
-        if self._waypoints:
-            w.write('waypoints: [')
-            w.indent()
-            [w.write('{location: %s, stopover: false},' % waypoint) for waypoint in self._waypoints]
-            w.dedent()
-            w.write('],')
-        w.write('travelMode: "%s"' % self._travel_mode)
-        w.dedent()
-        w.write('''  
-            }, function(response, status) {
-                if (status == google.maps.DirectionsStatus.OK) {
-                    new google.maps.DirectionsRenderer({map: map}).setDirections(response);
-                }
-            });
-        ''')
-        w.write()
-
-class _Text(object):
-    '''For more info, see Google Maps' `Directions Service https://developers.google.com/maps/documentation/javascript/directions`_.'''
-    
-    def __init__(self, lat, lng, text, **kwargs):
-        '''
-        Args:
-            lat (float): Latitude of the text label.
-            lng (float): Longitude of the text label.
-            text (str): Text to display.
-
-        Optional:
-
-        Args:
-            color/c (str): Text color. Can be hex ('#00FFFF'), named ('cyan'), or matplotlib-like ('c'). Defaults to black.
-        '''
-        self._position = _format_LatLng(lat, lng)
-        self._text = text
-        self._color = _get_hex_color(_get_value(kwargs, ['color', 'c'], '#000000'))
-        self._icon = _get_embeddable_image(_COLOR_ICON_PATH % 'clear')
-
-    def write(self, w):
-        '''
-        Write the text.
-
-        Args:
-            w (_Writer): Writer used to write the text.
-        '''
-        w.write('new google.maps.Marker({')
-        w.indent()
-        w.write('label: {')
-        w.indent()
-        w.write('text: "%s",' % self._text)
-        w.write('color: "%s",' % self._color)
-        w.write('fontWeight: "bold"')
-        w.dedent()
-        w.write('},')
-        w.write('icon: "%s",' % self._icon)
-        w.write('position: %s,' % self._position)
-        w.write('map: map')
-        w.dedent()
-        w.write('});')
-        w.write()
-
-class _MarkerIcon(object):
-    def __init__(self, color):
-        '''
-        Args:
-            color (str): Color of the marker icon.
-        '''
-        self._color = _get_hex_color(color)
-
-        self.name = 'marker_icon_%s' % self._color[1:]
-        '''str: JavaScript name of the marker icon.'''
-
-        # Get this marker icon as an embeddable image:
-        get_marker_icon_path = lambda color: _COLOR_ICON_PATH % color[1:]
-        marker_icon_path = get_marker_icon_path(self._color)
-
-        if not os.path.exists(marker_icon_path):
-            warnings.warn(" Marker color '%s' isn't supported." % self._color)
-            marker_icon_path = get_marker_icon_path('#000000')
-
-        self._icon = _get_embeddable_image(marker_icon_path)
-
-    def write(self, w, color_cache):
-        '''
-        Write the marker icon (if it isn't already written).
-
-        Args:
-            w (_Writer): Writer used to write the marker icon.
-            color_cache (set): Cache of colors written so far.
-        '''
-        # If this marker icon hasn't been written before, then embed it in the script:
-        if self._color not in color_cache:
-            w.write('var %s = {' % self.name)
-            w.indent()
-            w.write('url: "%s",' % self._icon) 
-            w.write('labelOrigin: new google.maps.Point(10, 11)') # TODO: Avoid hardcoded label origin.
-            w.dedent()
-            w.write('};')
-            w.write()
-            color_cache.add(self._color)
-
-class _Marker(object):
-    def __init__(self, position, **kwargs):
-        '''
-        Args:
-            position (str): JavaScript code that represents the position of the marker.
-
-        Optional:
-
-        Args:
-            name (str): JavaScript name of the marker.
-            title (str): Hover-over title of the marker.
-            label (str): Label displayed on the marker.
-            icon (str): JavaScript code that represents the icon.
-            draggable (bool): Whether or not the marker is draggable.
-        '''
-        self._position = position
-        self._name = kwargs.get('name')
-        self._title = kwargs.get('title')
-        self._label = kwargs.get('label')
-        self._icon = kwargs.get('icon')
-        self._draggable = kwargs.get('draggable')
-
-    def write(self, w):
-        '''
-        Write the marker.
-
-        Args:
-            w (_Writer): Writer used to write the marker.
-        '''
-        if self._name is not None: w.write('var %s = ' % self._name, end_in_newline=False)
-
-        w.write('new google.maps.Marker({')
-        w.indent()
-        w.write('position: %s,' % self._position)
-
-        if self._title is not None: w.write('title: "%s",' % self._title)
-        if self._label is not None: w.write('label: "%s",' % self._label)
-        if self._icon is not None: w.write('icon: %s,' % self._icon)
-        if self._draggable is not None: w.write('draggable: %s,' % str(self._draggable).lower())
-
-        w.write('map: map')
-        w.dedent()
-        w.write('});')
-        w.write()
-
-class _MarkerDropper(object):
-    '''
-    Handler that drops markers on map clicks.
-
-    The markers can be deleted when clicked on.
-    '''
-    _MARKER_NAME = 'dropped_marker'
-    _EVENT_OBJECT_NAME = 'event'
-
-    def __init__(self, color, **kwargs):
-        '''
-        Args:
-            color (str): Color of the markers to be dropped.
-
-        Optional:
-
-        Args:
-            title (str): Hover-over title of the markers to be dropped.
-            label (str): Label displayed on the markers to be dropped.
-            draggable (bool): Whether or not the markers to be dropped are draggable.
-        '''
-        self._marker_icon = _MarkerIcon(color)
-        self._marker = _Marker(
-            '%s.latLng' % self._EVENT_OBJECT_NAME,
-            name=self._MARKER_NAME,
-            title=kwargs.get('title'),
-            label=kwargs.get('label'),
-            icon=self._marker_icon.name,
-            draggable=kwargs.get('draggable')
-        )
-
-    def write(self, w, color_cache):
-        '''
-        Write the marker dropper.
-
-        Args:
-            w (_Writer): Writer used to write the marker dropper.
-            color_cache (set): Cache of colors written so far.
-        '''
-        # Write the marker icon (if it isn't written already).
-        self._marker_icon.write(w, color_cache)
-
-        # Write the marker-dropping handler:
-        w.write('map.addListener("click", function(%s) {' % self._EVENT_OBJECT_NAME)
-        w.indent()
-        self._marker.write(w)
-        w.write('''
-        {marker_name}.addListener('click', function() {{
-            {marker_name}.setMap(null);
-        }});
-        '''.format(marker_name=self._MARKER_NAME))
-        w.dedent()
-        w.write('});')
-        w.write()
-
-class _MarkerInfoWindow(object):
-    def __init__(self, marker_name, content):
-        '''
-        Args:
-            marker_name (str): JavaScript name of the marker that should display this info window.
-            content (str): HTML content to be displayed in this info window.
-        '''
-        self._marker_name = marker_name
-        self._content = content
-
-    def write(self, w, info_marker_index):
-        '''
-        Write the info window that attaches to the given marker on click.
-
-        Args:
-            w (_Writer): Writer used to write the info window.
-            info_marker_index (int): Index of this info window.
-        '''
-        w.write('''
-        var {info_window_name} = new google.maps.InfoWindow({{
-            content: '{content}'
-        }});
-
-        {marker_name}.addListener('click', function() {{
-            {info_window_name}.open(map, {marker_name});
-        }});
-        '''.format(
-            info_window_name='info_window_%d' % info_marker_index,
-            marker_name=self._marker_name,
-            content=self._content.replace("'", "\\'").replace("\n", "\\n") # (escape single quotes and newlines)
-        ))
-        w.write()
 
 class GoogleMapPlotter(object):
     '''
